lv_group: Rework event data
A key can be sent through `lv_group_send_key` which will create a fake
event for the key.

This allows previous uses of `lv_group_send_data` to be replaced
as-they-were.

The new parameter is the *whole* input of the keyboard event. This
allows us to get the string for the input. This is required for properly
handling UTF-8 input sequences *that may be more than a single
character*.