drm: Work off the first working connector
This is because we support only one display, and in my experience
platforms with multiple connectors work better off the first connector
rather than the following ones.

Here it's razer-cheryl2, where the second connector is (assumed to be)
the eDP output. It's not fine to assume a user boots with the eDP output
connected.