Revert "drm: Implement double-buffering"
This reverts commit 4372db750fed43b9d009a30451ce475fabbe29a3.

We're reverting for the time being; cheating with one buffer works
better in the end for speed. I prefer speedy updates compared to
reducing tearing.

I'm keeping this as a commit/revert pair in the history to allow future
programmers (maybe me!) to start from a known working implementation of
double-buffering.