lv_indev: Add event type to event data
This fixes an issue where event handlers tried to interpret the data
even though it wasn't the right type.

The "press" state is *sigh* shared between keyboard and touchscreen.

This meant that the handled for keyboard worked on bogus data.

What's confusing is how it works without issue on x86_64, verified to
work on combined data types. Even more confusing, how it doesn't cause
an issue when running the `examples/hello` system, nor the recovery
menu, but only in the splash, and only on some device types.