lv_indev: Fix typing fast with a keyboard
The issue was masked by the previous deficiencies. It was assumed that
future events shouldn't propagate until a key was entirely released.
This is problematic with fast typers, as it's entirely possible to start
pressing another key while releasing, which would be dropped here.

With this change, we're back to a proper typing experience.

What this removes is the implementation of "long pressing" using the
keyboard. Which is not desirable with our use case.