drm: Try all cards
On some platforms, e.g. MT8183, the output GPU is not card0. This failed
in a way the `fbdev` driver took over, but this means that the intrinsic
orientation of the display is not respected.

With this change, all cards will be tried to be used, before then
falling back.