lv_canvas_set_buffer: Use lv_img_dsc_t*
This makes the API symmetric with `lv_img_buf_alloc`.

For our use-case, it's fine since we're not intending to use manually
managed buffers or "program data" as a buffer. We're working on top of a
full operating system.

If dealing with a buffer manually ends-up being needed, we can add some
escape hatch.